+++ example/lib/main.dart
@@ -116,6 +116,7 @@ class _HomePageState extends State<HomePage> {
                   context: context,
                   initialTime: IRTimeModel(hour: 18, minute: 45, second: 59),
                   title: "انتخاب زمان",
+                  visibleSecondsPicker: true,
                   visibleNowButton: true,
                   nowButtonText: "انتخاب اکنون",
                   confirmButtonText: "تایید",